server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart  # for file uploads

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq
langchain-pinecone

# Vectorstore
# chromadb
pinecone

# Embeddings
# sentence-transformers
# langchain-google-generative-ai
langchain-google-genai

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm

# Logging (optional but recommended)
loguru



#  uv pip install -r requirements.txt